\name{AML43}
\alias{AML43}
\docType{data}
\title{
    SNA input for primary tumor and relapse genome of leukemia patient from
    Ding et al. Nature 2012.
}
\description{
    1242 SNAs from sequencing of leukemia patient at two timepoints. All SNAs
    are filtered to be from copy-number-neutral region.
}
\usage{data(AML43)}
\value{
    List of simulated SNA input data for Canopy.
}
\author{
    Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
    data(AML43)
}
\keyword{datasets}