\name{MDA231}
\alias{MDA231}
\docType{data}
\title{
Dataset for project MDA231
}
\description{
Pre-stored dataset for project MDA231. A transplantable metastasis model
system was derived from a heterogeneous human breast cancer cell line 
MDA-MB-231. Cancer cells from the parental line MDA-MB-231 were engrafted into
mouse hosts leading to organ-specific metastasis. Mixed cell populations (MCPs)
were in vivo selected from either bone or lung metastasis and grew into 
phenotypically stable and metastatically competent cancer cell lines. The
parental line as well as the MCP sublines were whole-exome sequenced with 
somatic SNAs and CNAs profiled.
}
\usage{data(MDA231)}
\value{
List of input data for Canopy from project MDA231.
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(MDA231)
}
\keyword{datasets}