\name{MDA231_sampchain}
\alias{MDA231_sampchain}
\docType{data}
\title{
List of pre-sampled trees
}
\description{
List of sampleed trees in subtree space with different number of subclones for
project MDA231.
}
\usage{data(MDA231_sampchain)}
\value{
List of sampled trees from different subtree space
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(MDA231_sampchain)
}
\keyword{datasets}