\name{MDA231_tree}
\alias{MDA231_tree}
\docType{data}
\title{
Most likely tree from project MDA231
}
\description{
Most likely tree from project MDA231 as a tree example.
}
\usage{data(MDA231_tree)}
\value{
Most likely tree from project MDA231
}
\author{
Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
data(MDA231_tree)
}
\keyword{datasets}