\name{toy3}
\alias{toy3}
\docType{data}
\title{
    Toy dataset 3 for Canopy
}
\description{
    Pre-stored simulated toy dataset. 200 simulated SNAs from a tree with 4
    branches. No CNA events at play.
}
\usage{data(toy3)}
\value{
    List of simulated SNA input data for Canopy.
}
\author{
    Yuchao Jiang \email{yuchaoj@wharton.upenn.edu}
}
\examples{
    data(toy3)
}
\keyword{datasets}